### 2023-08-24 #5780 @editorialbot
[PRE REVIEW]: mathlib: A Scala package for readable, verifiable and sustainable simulations of formal theory
**Submitting author:** @markblokpoel (<a href="http://orcid.org/0000-0002-1522-0343">Mark Blokpoel</a>)
**Repository:** https://github.com/markblokpoel/mathlib
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.0
