### 2023-08-29 #5780 @soodoku
> 👋 @soodoku - would you be willing to review this submission for JOSS?

I don't believe I have the right skills here to do a good job.
