### 2023-09-02 #5780 @danielskatz
👋 @mahfuz05062, @larkz, @jjerphan, @neurons - would a couple of you be willing to review this JOSS submission?
