### 2023-09-04 #5780 @jjerphan
Hi @danielskatz, thanks for proposing, unfortunately I do not have any bandwidth for reviewing submissions to JOSS.
