### 2023-09-15 #5780 @danielskatz
👋 @darrenjw and @romac - Would one of you be willing to review this submission ([short paper](https://github.com/openjournals/joss-papers/blob/joss.05780/joss.05780/10.21105.joss.05780.pdf) and [software](https://github.com/markblokpoel/mathlib)) for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
