### 2023-09-15 #5780 @darrenjw
I've read the paper and spent half an hour looking at the code repositories. But I still have no clue what it is, what it does, or what problem it is trying to solve. This suggests that either I don't have the expertise to review it, or that neither the paper nor the software are ready for formal review. Possibly both. I can attempt to review it if you like, but I'm not sure how helpful it will be.
