### 2023-09-15 #5780 @darrenjw
Actually, having now skimmed the associated online textbook, I think that the author may have re-discovered *property-based testing*. So my advice to the author would be to carefully study existing property-based testing frameworks for Scala, such as ScalaCheck - https://scalacheck.org/
