### 2023-09-19 #5780 @bzz
> I've read the paper and spent half an hour looking at the code repositories. But I still have no clue what it is, what it does, or what problem it is trying to solve.

After a cursory glance I had to admit beeing a bit confused. But I’m not an expert on cognitive science or psychology.
Would be curious to know if looking at https://computationalcognitivescience.github.io/lovelace/part_i/intro helps to clarify the library purpose/applications?
