### 2023-11-10 #5780 @markblokpoel
> Actually, having now skimmed the associated online textbook, I think that the author may have re-discovered _property-based testing_. So my advice to the author would be to carefully study existing property-based testing frameworks for Scala, such as ScalaCheck - https://scalacheck.org/

@danielskatz sorry for taking so long to reply. With respect to the comment above, it’s not property-based testing. The library has a different goal, viz., to facilitate writing code that is provably correct given a formal specification.
