### 2023-11-10 #5780 @danielskatz
> @danielskatz many thanks for nominating me, I'd be happy to review this. My duties are a bit full this semester, but how does a completion date around mid-November/early December sound?

@larkz - given the current date, I'm going to accept your offer, and assign you, and start the review
