### 2023-11-21 #6049 @danielskatz
👋 @bzz and @larkz - Just a brief reminder that you will need to use the command `@editorialbot generate my checklist` to create your review checklist before you can do anything else. @editorialbot commands need to be the first thing in a new comment.
