### 2023-11-22 46fc6cf
Set sum bugfix
Added further test cases and implemented fix by transforming Set to List to prevent duplicate numbers from disappearing after a map().
