### 2023-12-01 39bd5cf
Draft barabasiAlbertGraph
Also had to update probabilities because of scala version bump.
