+++ src/main/scala/mathlib/graph/UnDiGraph.scala
@@ -1,6 +1,6 @@
 package mathlib.graph
 
-import mathlib.graph.GraphImplicits.N
+import mathlib.graph.GraphImplicits.{EdgeImpl, N}
 import mathlib.set.SetTheory.ImplSet
 
 import scala.annotation.tailrec
@@ -54,6 +54,12 @@ case class UnDiGraph[T](
       val v: Set[Node[T]] = this.vertices union thatVertices
       val e               = this.edges union thatEdges
       UnDiGraph(v, e)
+  }
+
+  def toDOTString: String = {
+    "graph G {\n" +
+      edges.map(edge => "\t" + edge.left.label + " -- " + edge.right.label).mkString("\n") +
+    "\n}"
   }
 }
 
@@ -213,5 +219,67 @@ case object UnDiGraph {
   def uniform(n: Int, numberEdges: Int): UnDiGraph[String] = {
     val objects = (0 to n).toSet.map("N" + _)
     uniform(objects, numberEdges)
+  }
+
+  def barabasiAlbertGraph[T](size: Int, m: Int): UnDiGraph[String] = {
+    @tailrec
+    def barabasiAlbertGraph(n: Int, partialGraph: UnDiGraph[String]): UnDiGraph[String] = {
+      if(n + m == size) partialGraph
+      else {
+        val degrees = partialGraph.adjacencyList
+          .mapValues(_.size) // Get degrees of all vertices
+        val sumDegrees = degrees.values.sum
+        val probabilities = degrees.mapValues(_.doubleValue() / sumDegrees)
+        val rnd = Random.nextDouble()
+
+        @tailrec
+        def sampleVertices(
+          currentVertex: Node[String],
+          s: Int,
+          currentP: Double,
+          nextVertices: Seq[Node[String]],
+          sampledVertices: Seq[Node[String]]
+        ): Seq[Node[String]] = {
+          if (s == 0) sampledVertices
+          else {
+            if (nextVertices.isEmpty || currentP + probabilities(nextVertices.head) > rnd) {
+              val verticesWithoutCurrentVertex = partialGraph.vertices - currentVertex
+              sampleVertices(
+                verticesWithoutCurrentVertex.toSeq.head,
+                s - 1,
+                0.0,
+                verticesWithoutCurrentVertex.toSeq.tail,
+                sampledVertices :+ currentVertex
+              )
+            } else {
+              sampleVertices(
+                nextVertices.head,
+                s,
+                currentP + probabilities(nextVertices.head),
+                nextVertices.tail,
+                sampledVertices
+              )
+            }
+          }
+        }
+        val nodes = partialGraph.vertices.toSeq
+
+        val sampledVertices = sampleVertices(
+          nodes.head,
+          m,
+          0,
+          nodes.tail,
+          Seq.empty
+        )
+
+        val newVertex = N("V" + (n + m))
+        val newEdges = sampledVertices.map(sv => newVertex ~ sv).toSet
+        barabasiAlbertGraph(n + 1, partialGraph + newEdges)
+      }
+    }
+
+    val initialVertices = (0 until m).toSet.map((i: Int) => N("V" + i))  // Construct m vertices
+    val initialGraph =  UnDiGraph[String](initialVertices)
+    barabasiAlbertGraph(m, initialGraph)
   }
 }
