### 2023-12-04 #6049 @danielskatz
👋 @bzz and @larkz - I would appreciate it if you would use the command `@editorialbot generate my checklist` to create your review checklist, and then check off the first two items to make sure the process works correctly. @editorialbot commands need to be the first thing in a new comment.
