### 2023-12-08 c39b4cb
Renamed barabasiAlbertGraph to preferentialAttachment.
