### 2023-12-08 49e0060
Implemented preferential attachement for all basic graph types
