+++ src/main/scala/mathlib/graph/WDiGraph.scala
@@ -53,6 +53,12 @@ case class WDiGraph[T](
       val e               = this.edges union thatEdges
       WDiGraph(v, e)
   }
+
+  override def toDOTString: String = {
+    "graph G {\n" +
+      edges.map(edge => "\t" + edge.left.label + " -> " + edge.right.label + " [weight="+edge.weight+"]").mkString("\n") +
+      "\n}"
+  }
 }
 
 /** Factory for weighted directed graphs. */
@@ -215,5 +221,25 @@ case object WDiGraph {
   def uniform(n: Int, numberEdges: Int): WDiGraph[String] = {
     val objects = (0 to n).toSet.map("N" + _)
     uniform(objects, numberEdges)
+  }
+
+  /** Generates a weighted directed graph of size vertices using preferential attachment,
+   * first decsribed by Eggenberger and Pólya and also known as a Barabasi-Albert graph.
+   * Each edge will have a weight between 0 and upperbound.
+   *
+   * Eggenberger, F. & Pólya, G. J. Appl. Math. Mech. (ZAMM) 3, 279–289 (1923).
+   *
+   * Barabási, Albert-László; Albert, Réka (October 1999). "Emergence of scaling in
+   * random networks" (PDF). Science. 286 (5439): 509–512.
+   *
+   * @param size       The size of the generated network.
+   * @param m          The number initial vertices.
+   * @param upperbound The upperbound for randomly generated weights, defaults to 1.0.
+   * @return An undirected graph.
+   */
+  def preferentialAttachment(size: Int, m: Int, upperbound: Double = 1.0): WDiGraph[String] = {
+    val unweightedGraph = DiGraph.preferentialAttachment(size, m)
+    val weightedEdges = unweightedGraph.edges.map(e => WDiEdge(e.left, e.right, Random.nextDouble() * upperbound))
+    WDiGraph(unweightedGraph.vertices, weightedEdges)
   }
 }
