### 2023-12-18 #6049 @larkz
Thank you so much for your effort in providing this very interesting piece of code.

**Disclaimer:** This review is provided by an individual fairly experienced in Scala and JVM development, along with having a good grasp of probability theory. While I have rudimentary knowledge of set theory and graph theory, I am not super-familiar with  'formal theory' or prior work in Blokpoel, M., & van Rooij, I. (2021).

**Lines 6-9:**

Regarding the phrases, "These methods allow scientists to ‘conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined…" and "…They make otherwise underspecified theories precise and open for critical reflection."

This wording is rather vague, and it could apply to anything. For instance:

- "(Finite element analysis) can enable scientists to ‘conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined’."
- "(Stochastic calculus) can allow scientists to ‘conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined’."

It would be beneficial to provide a more specific explanation of what constitutes formal theory and offer concrete examples.

**Line 14:**

It is recommended to omit "cannot (easily)" for precision. If a problem cannot be solved, the predictions are not "derived" but rather "approximated." Additionally, defining the term "easily" is essential – does it refer to computational time savings, memory efficiency, or ease of problem formulation for coders?

**Lines 15-20:**

The author contends that Scala is an ideal language for modelling mathematical expressions. Concrete examples, especially for those unfamiliar with Scala, would enhance the clarity of this assertion.

**Line 25:**

While the claim about backwards compatibility is generally valid, it would be beneficial to elaborate on how this constitutes a tremendous advantage over other programming languages. Backwards compatibility is a common goal for many languages/frameworks; is mathlib notably more robust in this aspect compared to others?

**Lines 28-27:**

The argument seems to be that Scala's functional programming nature, resembling mathematical language, is its primary advantage. Is this the sole advantage, and are there any trade-offs in terms of runtime? A comparison with a non-functional language like Java, highlighting concrete improvements in Scala, would provide a stronger stance.

**Lines 46-47:**

Again, similar to Line 25, a more nuanced discussion about the uniqueness of backwards compatibility in Scala compared to other languages/frameworks would be beneficial.

**Line 51:**

**Resources:**
There seems to be an issue accessing the editorial notebooks in all the provided links.
- [00.00-scala_preface.ipynb Preface](https://github.com/markblokpoel/mathlib-examples/blob/main/tutorial/00.00-scala_preface.ipynb)
- [01.[01-06]-scala_introduction.ipynb Introduction to Scala](https://github.com/markblokpoel/mathlib-examples/blob/main/tutorial/01.[01-06]-scala_introduction.ipynb)
- [00.[01-02]-set_theory.ipynb Coding set theory with mathlib](https://github.com/markblokpoel/mathlib-examples/blob/main/tutorial/00.[01-02]-set_theory.ipynb)

And for the ones that are “not yet available”, I would expect to see more of them become available, especially for the purposes of this review, so that I (and other users) can get a quicker grasp of how to use the framework, and its potential advantages. The repository should be complete work, with core examples, not work-in-progress.

**Additional Comments:**

Even though the paper is supposed to be concise, there is room for improvement in the clarity of the exposition. Enhancing the presentation would allow readers, potentially lacking expertise in formal theory, to quickly understand the novelty and technical contributions of the software, which currently remain unclear and challenging to ascertain.
