### 2023-12-22 #6049 @markblokpoel
@larkz thank you very much for the invaluable feedback. I'm out of office for the next two weeks, but will reply to your suggestions when I get back as soon as possible.

@danielskatz regarding the linked repository with the examples and tutorial: @larkz rightly points out these are still under development. I can fix the bugs with running the notebooks, but I perceived these as a separate academic contribution, next to the mathlib library (potentialy suitable when finished for publication in the sibling Journal of Open Source Education. Your  advise as editor would be most welcome.
