### 2024-01-08 #6049 @danielskatz
👋 @markblokpoel - sorry I missed this somehow...

> @danielskatz regarding the linked repository with the examples and tutorial: @larkz rightly points out these are still under development. I can fix the bugs with running the notebooks, but I perceived these as a separate academic contribution, next to the mathlib library (potentialy suitable when finished for publication in the sibling Journal of Open Source Education. Your advise as editor would be most welcome.

I think there are two potentially separable things:
1. As you suggest, examples and tutorials can be separate academic contributions
2. However, there is likely some minimum stable contribution needed as part of this submission.

So, perhaps you can separate and mark the work that is part of this submission, and that reviewers (and users) can rely on today to help them understand the software, and what is work-in-progress that will provide more support but isn't part of the minimum needed to start using the software.
