### 2024-02-28 abe7dc3
Revisions
- Rephrased "these methods"
- Explained 'easily' derived in a footnote.
