+++ paper.bib
@@ -16,6 +16,18 @@
     doi = {10.1177/1745691620970585}
 }
 
+@article{thagard:1998,
+    title = {Coherence as {{Constraint Satisfaction}}},
+    author = {Thagard, Paul and Verbeurgt, Karsten},
+    year = {1998},
+    journal = {Cognitive Science},
+    volume = {22},
+    number = {1},
+    pages = {24},
+    langid = {english},
+    keywords = {Coherence}
+}
+
 @article{vanrooij_baggio:2021,
     url = {https://journals.sagepub.com/doi/full/10.1177/1745691620970604},
     author = {{van Rooij}, I. and {Baggio}, G.},
