+++ paper.md
@@ -94,17 +94,22 @@ def subsetChoice(
 }
 ```
 
-| Formal expression                     | ```mathlib implementation```               | Description                                                                                                             |
-|---------------------------------------|--------------------------------------------|-------------------------------------------------------------------------------------------------------------------------|
-| n.a.                                  | ```Item```                                 | Custom type for items.                                                                                                  |
-| $I$                                   | ```items: Set[Item]```                     | A set of items.                                                                                                         |
-| $v:I\rightarrow \mathbb{Z}$           | ```v: (Item => Double)```                  | Value function for single items.                                                                                        |
-| $b:I \times I \rightarrow \mathbb{Z}$ | ```b: ((Item, Item) => Double)```          | Value function for pairs of items.                                                                                      |
-| n.a.                                  | ```def value(subset: Set[Item]): Double``` | Function wrapper for the combined value of a subset.                                                                    |
-| $\sum_{i \in I'}v(i)$                 | ```sum(subset, v)```                       | Sum of single item values, where ```subset``` is $I$.                                                                   |
-| $\sum_{i, j \in I'}b(i,j)$            | ```sum(subset.uniquePairs, b)```           | Sum of pair-wise item values, where ```uniquePairs``` generates all pairs ```(x, y)``` in ```subset``` with ```x!=y```. |
-|                                       |                                            |                                                                                                                         |
-|                                       |                                            |                                                                                                                         |
+| Formal expression                     | ```mathlib``` implementation and description                                                                            |
+|---------------------------------------|-------------------------------------------------------------------------------------------------------------------------|
+| n.a.                                  | ```Item```                                                                                                              |
+|                                       | Custom type for items.                                                                                                  |
+| $I$                                   | ```items: Set[Item]```                                                                                                  |
+|                                       | A set of items.                                                                                                         |
+| $v:I\rightarrow \mathbb{Z}$           | ```v: (Item => Double)```                                                                                               | 
+|                                       | Value function for single items.                                                                                        |
+| $b:I \times I \rightarrow \mathbb{Z}$ | ```b: ((Item, Item) => Double)```                                                                                       | 
+|                                       | Value function for pairs of items.                                                                                      |
+| n.a.                                  | ```def value(subset: Set[Item]): Double```                                                                              | 
+|                                       | Function wrapper for the combined value of a subset.                                                                    |
+| $\sum_{i \in I'}v(i)$                 | ```sum(subset, v)```                                                                                                    | 
+|                                       | Sum of single item values, where ```subset``` is $I$.                                                                   |
+| $\sum_{i, j \in I'}b(i,j)$            | ```sum(subset.uniquePairs, b)```                                                                                        | 
+|                                       | Sum of pair-wise item values, where ```uniquePairs``` generates all pairs ```(x, y)``` in ```subset``` with ```x!=y```. |
 
 
 ## Illustration 2: Coherence
@@ -115,7 +120,7 @@ def subsetChoice(
 
 *Input:* A graph $G=(V,E)$ with vertex set $V$ and edge set $E\subseteq V\times V$ that partitions into positive constraints $C^+$ and negative constraints $C^-$ (i.e., $C^+\cup C^-=E$ and $C^+\cap C^-=\varnothing$) and a weight function $w: E \rightarrow \mathbb{R}$.
 
-*Output:* A truth value assignemnt $T:V \rightarrow \{true, false\}$ such that $Coh(T)=Coh^+(A,R)+Coh^-(T)$ is maximum. Here,
+*Output:* A truth value assignment $T:V \rightarrow \{true, false\}$ such that $Coh(T)=Coh^+(A,R)+Coh^-(T)$ is maximum. Here,
 $$
 Coh^+(T)=\displaystyle\sum_{(u,v)\in C^+}
 \begin{cases}
