+++ paper.md
@@ -34,7 +34,7 @@ be analytically derived[^1].
 thought to analytically derive, even when in principle the property can be analytically derived.
 
 ```mathlib``` is a library for Scala [@odersky:2008] supporting functional programming that resembles mathematical 
-expressions such as set theory, graph theory and probability theory. This library was developed to complement the
+expressions such as set theory and graph theory. This library was developed to complement the
 theory development method outlined in the open education book Theoretical modeling for cognitive science and
 psychology by @blokpoel_vanrooij:2021.
 
@@ -47,33 +47,35 @@ Scala using ```mathlib``` is:
 
 # Statement of need
 
-Writing code is not easy, writing code for which we can know that it computes what the specification
-(i.e., the formal theory) states is even harder. This can be facilitated by having a programming language where the
-syntax and semantics closely matches that of the specification. Since formal theories are specified using mathematical
-notation [@marr:1982, @blokpoel_vanrooij:2021, @guest_martin:2021], functional programming languages bring a lot to the
-table in terms of syntactic and semantic resemblance to mathematical concepts and notation. ```mathlib``` adds
-mathematical concepts and notation to the functional programming language Scala [@odersky:2008], specifically in the
-current version (0.9.1): set theory and graph theory. ```mathlib``` differs from other libraries in that it focuses on
-usability and transparency, whereas other libraries focus on computational expressiveness and efficiency.
-
+```mathlib``` supports scholars to write **readable** and **verifiable** code. Writing code is not easy, writing
+code for which we can know that it computes what the specification (i.e., the formal theory) states is even harder.
 Given the critical role of theory and computer simulations in cognitive science, it is important that scholars
-can verify that the code does what the authors intent it to do. We provide two examples to illustrate how Scala 
-and ```mathlib``` make code more accessible to verify the relationship between simulation and theory.
+can verify that the code does what the authors intent it to do. This can be facilitated by having a programming language
+where the
+syntax and semantics closely matches that of the specification. Since formal theories are specified using mathematical
+notation [@marr:1982; @blokpoel_vanrooij:2021; @guest_martin:2021], functional programming languages bring a lot to the
+table in terms of syntactic and semantic resemblance to mathematical concepts and notation. ```mathlib``` adds
+mathematical concepts and notation to the functional programming language Scala [@odersky:2008]. At the time of writing,
+the current version (0.9.1) supports set theory and graph theory. In the next section, we provide two examples to
+illustrate how Scala and ```mathlib``` make code more accessible to verify the relationship between simulation and
+theory.
 
-```mathlib``` and Scala support scholars to address the issue of out-of-date code. It is important that
+```mathlib``` and Scala support scholars write **sustainable** code. It is important that
 academic contributions remain accessible for reflection and critique. This includes simulations that also have
 theoretical import. Simulation results may need to be verified, or future scholars may wish to expand upon the work.
 It is not sufficient to archive code, because in practise programming contributions in academia are easily lost because
-of incompatibility issues between older software and newer operating systems. Scala (and consequently ```mathlib```) 
+of incompatibility issues between older software and newer operating systems. Scala (and consequently ```mathlib```)
 runs on the Java Virtual Machine (JVM) and has state-of-the-art versioning. The programmer can specify exactly which
 version of Scala and ```mathlib``` needs to be retrieved to run the code on any system that supports the JVM. Even
 when newer versions of Scala or ```mathlib``` may potentially break older code, this versioning system allows future
 users to easily run older code.
 
-What makes ```mathlib``` unique is that its design encourages the computational cognitive scientist to write
[…]
