+++ src/main/scala/mathlib/demos/Coherence.scala
@@ -6,101 +6,40 @@ import mathlib.set.SetTheory._
 
 object Coherence {
 
-  def main(args: Array[String]): Unit = {
+  def coherence(
+      network: WUnDiGraph[String],
+      positiveConstraints: Set[WUnDiEdge[Node[String]]]
+  ): Map[Node[String], Boolean] = {
+    val negativeConstraints: Set[WUnDiEdge[Node[String]]] =
+      network.edges \ positiveConstraints
 
-    def isPositiveConstraintSatisfied(
-        positiveConstraint: WUnDiEdge[Node[String]],
-        assignment: Map[Node[String], Boolean]
-    ): Boolean = assignment(positiveConstraint.left) == assignment(positiveConstraint.right)
+    def cohPlus(assignment: Map[Node[String], Boolean]): Double = {
+      def isSatisfied(pc: WUnDiEdge[Node[String]]): Double =
+        if (assignment(pc.left) == assignment(pc.right)) pc.weight
+        else 0.0
 
-    def isNegativeConstraintSatisfied(
-        negativeConstraint: WUnDiEdge[Node[String]],
-        assignment: Map[Node[String], Boolean]
-    ): Boolean = assignment(negativeConstraint.left) != assignment(negativeConstraint.right)
-
-    def cohPlus(
-        assignment: Map[Node[String], Boolean],
-        positiveConstraints: Set[WUnDiEdge[Node[String]]]
-    ): Double =
-      positiveConstraints.toList
-        .map((pc: WUnDiEdge[Node[String]]) => {
-          if (isPositiveConstraintSatisfied(pc, assignment)) pc.weight
-          else 0.0
-        })
-        .sum
-
-    def cohMinus(
-        assignment: Map[Node[String], Boolean],
-        negativeConstraints: Set[WUnDiEdge[Node[String]]]
-    ): Double =
-      negativeConstraints.toList
-        .map((nc: WUnDiEdge[Node[String]]) => {
-          if (isPositiveConstraintSatisfied(nc, assignment)) nc.weight
-          else 0.0
-        })
-        .sum
-
-    def coh(
-        assignment: Map[Node[String], Boolean],
-        positiveConstraints: Set[WUnDiEdge[Node[String]]],
-        negativeConstraints: Set[WUnDiEdge[Node[String]]]
-    ): Double =
-      cohPlus(assignment, positiveConstraints) + cohMinus(assignment, negativeConstraints)
-
-//    def cMinusCoherence(
-//        network: WUnDiGraph[String],
-//        positiveConstraints: Set[WUnDiEdge[Node[String]]],
-//        negativeConstraints: Set[WUnDiEdge[Node[String]]]
-//    ): Set[Map[Node[String], Boolean]] = {
-//      require(positiveConstraints \/ negativeConstraints == network.edges, "C+ union C- != E")
-//      require(
-//        positiveConstraints /\ negativeConstraints == Set.empty,
-//        "C+ intersect C- is not empty"
-//      )
-//
-//      def completeBeliefAssignment(
-//          negativeAssignments: Map[Node[String], Boolean]
-//      ): Map[Node[String], Boolean] = {
-//        val positiveAssignments = positiveConstraints
-//          .flatMap(pc => {
-//            if (negativeAssignments.contains(pc.left) && !negativeAssignments.contains(pc.right))
-//              Set(pc.right -> negativeAssignments(pc.left))
-//            else if (
-//              !negativeAssignments.contains(pc.left) && negativeAssignments.contains(pc.right)
-//            ) Set(pc.left -> negativeAssignments(pc.right))
-//            else if (
-//              !negativeAssignments.contains(pc.left) && !negativeAssignments.contains(pc.right)
-//            ) Set(pc.left -> true, pc.right -> true)
[…]
