### 2024-03-18 #6049 @markblokpoel
Dear @danielskatz and @larkz, thank you for your patience. I've completed my revisions. Please find the updated manuscript in the next post. Below I've replied to your points of feedback.

**Lines 6-9:**
Regarding the phrases, "These methods allow scientists to ‘conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined…" and "…They make otherwise underspecified theories precise and open for critical reflection."
This wording is rather vague, and it could apply to anything. For instance:
- "(Finite element analysis) can enable scientists to ‘conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined’."
- "(Stochastic calculus) can allow scientists to ‘conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined’."
It would be beneficial to provide a more specific explanation of what constitutes formal theory and offer concrete examples.

*__Reply:__* Thank you for pointing out this ambiguity. I have added clarification in the summary by giving examples of the formal systems. The approach in Blokpoel and van Rooij is agnostic about the specific formal systems that theoreticians can use, which is why I prefer to give examples here and not to be concrete. In addition to the subset choice illustration, I have also included an additional illustration (viz., Coherence) to hopefully give the reader a better idea what the method is about.

**Line 14:**
It is recommended to omit "cannot (easily)" for precision. If a problem cannot be solved, the predictions are not "derived" but  rather "approximated." Additionally, defining the term "easily" is  essential – does it refer to computational time savings, memory  efficiency, or ease of problem formulation for coders?

*__Reply:__* I agree that the word 'easily' is confusing. I did not intend approximation. In fact approximation has in our field of computational cognitive science problems in and of itself (see e.g., van Rooij & Wareham, 2012 and Kwisthout, Wareham & van Rooij, 2011). I intended the latter meaning, which I have hopefully clarified in Footnote 1.

- van Rooij, I. & Wareham, T. (2012). [Intractability and approximation of optimization theories of cognition](https://www.sciencedirect.com/science/article/pii/S002224961200048X). *Journal of Mathematical Psychology, 56*,232-247.
- Kwisthout, J., Wareham, T., & van Rooij, I. (2011). [Bayesian intractability is not an ailment that approximation can cure](http://onlinelibrary.wiley.com/doi/10.1111/j.1551-6709.2011.01182.x/pdf). *Cognitive Science, 35*(5), 779-784.

**Lines 15-20:**
The author contends that Scala is an ideal language for modelling mathematical expressions. Concrete examples, especially for those unfamiliar with Scala, would enhance the clarity of this assertion.

*__Reply:__* I have extended the illustration of Subset choice and added a second illustration on Coherence theory. I have also clarified which expressions in the code have been implemented in ``mathlib``. (Section **Illustrations**)

**Line 25:**
While the claim about backwards compatibility is generally valid, it would be beneficial to elaborate on how this constitutes a tremendous advantage over other programming languages. Backwards compatibility is a common goal for many languages/frameworks; is mathlib notably more robust in this aspect compared to others?
**Lines 46-47:**
Again, similar to Line 25, a more nuanced discussion about the uniqueness of backwards compatibility in Scala compared to other languages/frameworks would be beneficial.

*__Reply:__* True backwards compatibility is of course dependent on future iterations of the software. What is important for cognitive science, is that older code can easily be run and critically analyzed, courtesy of Maven and the Scala Built Tool. In cognitive science, it is very common that old simulations cannot be run or verified, because of the complications arising in running old software. I realize this is not backwards compatibility. I still belief that Scala has an edge on other programming languages, but it seems that this argument is outside the scope of this paper. I have limited the scope on my claims on backwards compatibility, which I've hopefully addressed in the paper (Lines 25; 40--50).

**Lines 28-27:**
The argument seems to be that Scala's functional programming nature, resembling mathematical language, is its primary advantage. Is this the sole advantage, and are there any trade-offs in terms of runtime? A comparison with a non-functional language like Java, highlighting concrete improvements in Scala, would provide a stronger
stance.

*__Reply:__* I agree that ideally a comparison might make the need for mathlib clearer, but at the same time I worry that such a comparison may lead to a discusion on *how one can also do functional programming in language X*, which I actually think is true. Also, I do not wish to argue that cognitive sciences *should* use mathlib, but the principles that it encourages are relevant for any theoretical simulation. This discussion, however, is outside the scope of the article. I did try to summarize the primary advantage as you so kindly summarized (lines 51-52), and I would like to point out that line 55-57 expresses the tradeoff we considered: usability and transparency versus computational expressiveness and efficiency.

**Line 51:** / **Resources:**
There seems to be an issue accessing the editorial notebooks in all the provided links.
- [00.00-scala_preface.ipynb Preface](https://github.com/markblokpoel/mathlib-examples/blob/main/tutorial/00.00-scala_preface.ipynb)
- [01.[01-06]-scala_introduction.ipynb Introduction to Scala](https://github.com/markblokpoel/mathlib-examples/blob/main/tutorial/01.%5B01-06%5D-scala_introduction.ipynb)
- [00.[01-02]-set_theory.ipynb Coding set theory with mathlib](https://github.com/markblokpoel/mathlib-examples/blob/main/tutorial/00.%5B01-02%5D-set_theory.ipynb)
- And for the ones that are “not yet available”, I would expect to see more of them become available, especially for the purposes of this review, so that I (and other users) can get a quicker grasp of how to use the framework, and its potential advantages. The repository should be complete work, with core examples, not work-in-progress.

*__Reply:__* The tutorials are work-in-progress to be submitted separately to the Journal of Open Source Education. I've removed the reference to them in the paper, as they are related to, but not part of this submission. I did update the demos in the mathlib repository for the two illustrations (Subset Choice and Coherence), and these should be runnable after checking out the mathlib repository and by using sbt to compile / run the project. References to these files have been added to the paper (lines 73-74 and lines 94-95).

**Additional Comments:**
Even though the paper is supposed to be concise, there is  room for improvement in the clarity of the exposition. Enhancing the  presentation would allow readers, potentially lacking expertise in  formal theory, to quickly understand the novelty and technical contributions of the software, which currently remain unclear and challenging to ascertain.

*__Reply:__* Thanks again for the review @larkz . I took the liberty of expanding the paper, I hope that still fits the journal constraints, @danielskatz ? If not, I'm happy to compress or otherwise update the paper.
