### 2024-03-22 b7e1d5e
Updated subset choice to yield only one solution.
