+++ paper.md
@@ -79,7 +79,8 @@ similar functionality focus on computational expressiveness and efficiency.
 
 # Illustrations
 
-We present illustrate the relationship between simulations implemented in Scala and ```mathlib``` and formal theories.
+We present two illustrations to show the relationship between simulations implemented in Scala
+and ```mathlib``` and formal theories.
 
 ## Illustration 1: Subset choice
 
@@ -107,12 +108,13 @@ def subsetChoice(
   items: Set[Item],
   v: Item => Double,
   b: (Item, Item) => Double
-): Set[Set[Item]] = {
+): Set[Item] = {
   
   def value(subset: Set[Item]): Double = 
     sum(subset, v) + sum(subset.uniquePairs, b)
 
-  argMax(powerset(items), value)
+  val allOptimalSolutions = argMax(powerset(items), value)
+  allOptimalSolutions.random.get
 }
 ```
 
