### 2024-05-06 #6049 @stephenfmann
I don't have experience with Scala but I do have a background in mathematics and some familiarity with van Rooij's work (from the perspective of philosophy of cognitive science). I'm happy to join if that's sufficient.
