### 2024-05-06 #6049 @drussellmrichie
I also don't have any experience with Scala but am a frequent (cognitive) modeler. How much time do we have to do the reviews? If I can have a few weeks, then I can do this.
