### 2024-05-06 #6049 @drussellmrichie
Hi Daniel. That will be fine.

On Mon, May 6, 2024 at 10:03 AM Daniel S. Katz ***@***.***>
wrote:

> @drussellmrichie <https://github.com/drussellmrichie> - I would be happy
> to add you as well. We aim to have the reviews done in ~4 weeks, which
> should work. Is this ok?
>
