### 2024-05-06 #6049 @danielskatz
@drussellmrichie - Thanks for agreeing to review this submission.
This is the review thread for the paper. All of our communications will happen here.

You should now use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
