### 2024-05-06 markblokpoel/mathlib#15 @stephenfmann
Typos in `paper.md`
I found some typos; I'm not sure if you would prefer me to correct them myself and make a pull request so I'm just listing them here instead:
+ "...that the code does what the authors *intent* it to do." => should be "intend"
+ "mathlib and Scala *support* scholars write sustainable code." => should be something like "encourage"? Or "support scholars *to* write sustainable code".
+ "in practise" => "in practice"
+ "that maximizes the combined value of the selected items *according*" => "accordingly"
+ There is a rogue underscore at the beginning of the final entry in Table 1.
+ "To *belief* that it rains and not wearing shorts is consistent, but to *belief* that it rains and to wear shorts is inconsistent." => belief should be "believe" both time.
+ In COHERENCE, you have a Coh+(A,R) that should presumably be Coh+(T).

Mentioning openjournals/joss-reviews#6049 for completeness.
