### 2024-05-07 #6049 @Nathaniel-Haines
> 👋 @Nathaniel-Haines - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @danielskatz, I also do not have any experience with Scala.  Given that both @stephenfmann and @drussellmrichie have agrred to review and also noted lack of experience with Scala, perhaps someone with Scala experience would be a better fit for my slot?
