### 2024-05-13 markblokpoel/mathlib#16 @stephenfmann
Some demo scripts fail
Some of the demo scripts are out of sync with release 0.9.0 and even 0.9.1, so for example `Coherence.scala` fails because `value preferentialAttachment is not a member of object mathlib.graph.WUnDiGraph`. An example of a demo script that works is `VertexCover.scala`.

Part of review https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2096068069
