### 2024-05-13 markblokpoel/mathlib#17 @stephenfmann
Error with `tutorial/02.02-set_theory-advanced_set_theory.ipynb`
I managed to install Jupyter Lab notebooks as documented in the [examples repository](https://github.com/markblokpoel/mathlib-examples). Most of the examples that have been written run as expected. However, attempting to run the second cell of `tutorial/02.02-set_theory-advanced_set_theory.ipynb`:
```
import $ivy.`com.github.markblokpoel.mathlib::mathlib:-SNAPSHOT
```
produces the following error:
```
Failed to resolve ivy dependencies:Error downloading com.github.markblokpoel.mathlib:mathlib_2.13:-SNAPSHOT
not found: ~\.ivy2\local\com.github.markblokpoel.mathlib\mathlib_2.13\-SNAPSHOT\ivys\ivy.xml
not found: https://repo1.maven.org/maven2/com/github/markblokpoel/mathlib/mathlib_2.13/-SNAPSHOT/mathlib_2.13--SNAPSHOT.pom
not found: https://jitpack.io/com/github/markblokpoel/mathlib/mathlib_2.13/-SNAPSHOT/mathlib_2.13--SNAPSHOT.pom
```

Part of https://github.com/openjournals/joss-reviews/issues/6049 (which is why I'm raising this issue here rather than in the examples repository).
