### 2024-05-13 markblokpoel/mathlib#18 @stephenfmann
Include link to documentation in README
By looking around I found https://markblokpoel.com/mathlib/scaladoc/mathlib/. This is useful and should be linked in the README.

Part of review https://github.com/openjournals/joss-reviews/issues/6049
