### 2024-05-13 #6049 @stephenfmann
Hi all, I've completed the first pass of my review. I have opened issues in the target repository. Here are some comments that don't really fit as issues:

+ The [documentation ](https://markblokpoel.com/mathlib/scaladoc/mathlib/)is generally good, but some of the objects don't have docstrings e.g. of the packages, only `set` has a docstring; it would be useful to have descriptions for the other packages too.
+ I couldn't find automated tests; at least some of the manual tests are failing as reported in issues https://github.com/markblokpoel/mathlib/issues/16 and https://github.com/markblokpoel/mathlib/issues/17
+ At least one of the reasons for those aforementioned failures appears to be a mismatch between the released version of `0.9.1` and the demo script in the current GitHub repository. But the README claims Scala code is "easy to sustain, because newer versions of Scala and mathlib can run old code (backwards compatibility)", so it would be useful to leverage this backwards compatibility to avoid this mismatch problem. (Since I'm not familiar with Scala I don't know how to do this without instruction.)
+ I haven't found information about **Community guidelines** yet (contributions and support); though the fact that the repository is public and on GitHub may suffice.
+ One of the checklist comments for the paper is "**State of the field**: Do the authors describe how this software compares to other commonly-used packages?" There's a discussion of Scala more generally but there don't appear to be other commonly-used packages for this specific use case. I'll leave it unchecked until the author confirms this is true.
