### 2024-05-13 #6049 @danielskatz
👋 @stephenfmann - thanks for your review and feedback. I'll let @markblokpoel respond to most of it, but I will say that the fact that the repository is public and on GitHub does not suffice as Community Guidelines. And thanks for opening issues and tagging them.
