### 2024-05-15 96c8c12
Update README.md
Added badge to link to scaladoc, see issue #18
