### 2024-05-15 markblokpoel/mathlib#18 @markblokpoel
I've added a badge to link to the scaladoc, thanks for pointing this out!
