### 2024-05-15 markblokpoel/mathlib#16 @markblokpoel
You are right that `Coherence.scala` demo uses a function that is currently only in beta version 0.9.2-b1. If you followed the installation instructions from the website, you probably used releases 0.9.0 or 0.9.1. These versions do not contain `preferentialAttachement` and `Coherence.scala` will subsequently not compile.

I will update the demo to use an older random graph generation function so it will compile under 0.9.1. This way the review process can move more efficiently. I'm aiming to release a new version 0.9.2 in the near future which will support more random graph algorithms.

A note on backwards compatibility, I've revised that statement in an [earlier revision](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2004659999). The idea is that code written in an earlier version can either be run using that older version (as you've tried), but newer code cannot be run with an older version of the library (this is the error you've encountered). I hope this clarifies things.
