### 2024-05-15 ed1b37d
Coherence compatibility v0.9.1
Removed useage of preferentialAttachement function which is not supported in latest non-beta version of mathlib, as reported in #16
