### 2024-05-15 markblokpoel/mathlib#16 @markblokpoel
I've updated the demo, can you confirm that it runs with your setup? Thanks!
