### 2024-05-15 markblokpoel/mathlib#17 @markblokpoel
Thank you for taking a look at the notebook tutorials. This is a peculiar error message. It lists jitpack which is an old code distribution method I used during pilot versions of mathlib. It is not referenced in the most recent versions. The import statement is also different from the one used in the latest version, which is:

```
`import $ivy.`com.markblokpoel::mathlib:0.9.0`
```

Can you confirm which version of the examples you downloaded?
