### 2024-05-15 markblokpoel/mathlib#17 @stephenfmann
Sure, I'll be able to take a look at this and https://github.com/markblokpoel/mathlib/issues/16#issuecomment-2111949780 this evening and will get back to you in both issue threads.
