### 2024-05-15 52f9fe2
Fixed typos
As kindly provided by @stephenfmann in #15.
