### 2024-05-15 markblokpoel/mathlib#15 @markblokpoel
Thank you so much for finding these typos! Great catch to find Coh+(A,R) which should indeed be Coh+(T). I was tempted to leave it in for other readers to discover, but I resisted. ;-)
