### 2024-05-15 #6049 @danielskatz
@markblokpoel - re community guidelines, some fairly fancy and complex versions are in https://github.com/corsa-center/oss-documents/blob/main/README.md#contributing-documents, but that document itself also has a fairly simple version at https://github.com/corsa-center/oss-documents/blob/main/README.md#about-this-repository

Another project I work on has https://github.com/Parsl/parsl?tab=readme-ov-file#contributing

But really all you have to do is meet the minimum criteria in https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines, which could be a sentence or so for each of the three points. You might want to look at some published JOSS papers and their associated software to see what they have done.
