### 2024-05-15 markblokpoel/mathlib#16 @stephenfmann
`Coherence.scala` is working; the others are all working except `GraphsDemo.scala` which throws the following on line 84:

```
value degreeOf is not a member of mathlib.graph.DiGraph[String]
    println(cycleCheck.degreeOf(N("a")))
```

This is for version `0.9.1`.
