### 2024-05-15 markblokpoel/mathlib#17 @stephenfmann
The README of the [mathlib-examples](https://github.com/markblokpoel/mathlib-examples) repository contains the line: "Finally, download a copy of the repository from the releases page and unpack the files." Under **Releases** on that repository there is just one link, with link text [1 tags](https://github.com/markblokpoel/mathlib-examples/tags). That tags page has one release that leads [here](https://github.com/markblokpoel/mathlib-examples/releases/tag/v0.1.0-beta), which is what I downloaded. I would suggest changing the README and/or the list of releases.

When I download the mathlib-examples repository's main branch as a zip, the tutorial notebooks look correct ([02.02-set_theory-advanced_set_theory.ipynb](https://github.com/markblokpoel/mathlib-examples/blob/v0.1.0-beta/notebooks/tutorial/02.02-set_theory-advanced_set_theory.ipynb) does indeed contain the updated line), with the exception of [03-graph-theory.ipynb](https://github.com/markblokpoel/mathlib-examples/blob/v0.1.0-beta/notebooks/tutorial/03-graph-theory.ipynb) which appears to be empty and [04-probability-theory.ipynb](https://github.com/markblokpoel/mathlib-examples/blob/v0.1.0-beta/notebooks/tutorial/04-probability-theory.ipynb) which appears to be unfinished. I would suggest either removing them from the main branch, or creating a release that omits them.

Similarly, the `notebooks/examples` repository contains some empty notebooks which I would remove.
