+++ README.md
@@ -9,13 +9,20 @@
 
 **This is a copy of the companion website main page, please visit [https://www.markblokpoel.com/mathlib](https://www.markblokpoel.com/mathlib).**
 
+[![mathlib Scala version support](https://index.scala-lang.org/markblokpoel/mathlib/mathlib/latest.svg)](https://index.scala-lang.org/markblokpoel/mathlib/mathlib)
+[![license](https://img.shields.io/badge/license-%20GPL--3.0-blue)](https://github.com/markblokpoel/mathlib/blob/master/LICENSE)
+[![Website](https://img.shields.io/website?url=https%3A%2F%2Fwww.markblokpoel.com%2Fmathlib)](https://www.markblokpoel.com/mathlib)
+![Maven Central](https://img.shields.io/maven-central/v/com.markblokpoel/mathlib_2.13)
+[![Scaladoc](https://img.shields.io/maven-central/v/com.markblokpoel/mathlib_2.13?label=scaladoc)](https://markblokpoel.com/mathlib/scaladoc/mathlib/)
+
 ```mathlib``` is a library for Scala supporting functional programming that resembles
 mathematical expressions such as set theory, graph theory and probability theory.
 This library was developed to complement the theory development method outlined
 in the open education book [Theoretical modeling for cognitive science and psychology by
 Blokpoel and van Rooij (2021)](https://computationalcognitivescience.github.io/lovelace/).
 
-The goal of this library is to facilitate users to implement simulations of their formal theories. ```mathlib``` and Scala code is:
+The goal of this library is to facilitate users to implement simulations of their formal theories. ```mathlib``` and
+Scala code is:
 
 * 👓 easy to **read**, because ```mathlib``` syntax closely resembles mathematical notation
 * ✅ easy to **verify**, by proving that the code exactly implements the theoretical model (or not)
@@ -27,28 +34,31 @@ For the scope of this documentation, we are going to assume you are familiar wit
 Some helpful links to get you started are:
 
 * [Using IntelliJ to run Scala applications in
-](https://www.jetbrains.com/help/idea/run-debug-and-test-scala.html#run_scala_app)
-* Use the binder service or a local instance of the Jupyterlab tutorials at [markblokpoel/mathlib-examples](https://github.com/markblokpoel/mathlib-examples)
+  ](https://www.jetbrains.com/help/idea/run-debug-and-test-scala.html#run_scala_app)
+* Use the binder service or a local instance of the Jupyterlab tutorials
+  at [markblokpoel/mathlib-examples](https://github.com/markblokpoel/mathlib-examples)
 
 ### Using this library with ```sbt```
+
 If you have set up a Scala project with ```sbt```, the Scala built tool, then add the following line to your
-```build.sbt``` file to enable the library. 
+```build.sbt``` file to enable the library.
 
 ```scala
-libraryDependencies += "com.markblokpoel" %% "mathlib" % "0.9.0"
+libraryDependencies += "com.markblokpoel" %% "mathlib" % "0.9.1"
 ```
 
 ### Using this library with Almond
+
 If you have set up a Scala project in jupyter notebooks using [Almond](https://almond.sh/) you can enable the library by
-adding the following line to a worksheet. 
+adding the following line to a worksheet.
 
 ```scala
-import $ivy.`com.markblokpoel::mathlib:0.9.0`
+import $ivy.`com.markblokpoel::mathlib:0.9.1`
 ```
 
 ### Import ```mathlib``` packages
 
-After enabling ```mathlib``` in your project you most likely will want to use some of its components. See the 
+After enabling ```mathlib``` in your project you most likely will want to use some of its components. See the
 [development section](#development) below for the available components and their import statements. Adding these
[…]
