### 2024-05-23 markblokpoel/mathlib#16 @markblokpoel
Thanks. I've realized that the `degreeOf` and `preferentialAttachement` functions are part of the 0.9.2-b1 release. I've removed these functions from `GraphsDemo.scala` until the release of 0.9.2 as I don't want to push for a hasty release of these new functions.
