### 2024-05-23 markblokpoel/mathlib#17 @markblokpoel
thank you for the suggestions. I've updated the repository and the release file. Its much cleaner this way, and hopefully should work fully under v0.9.1.
