### 2024-05-23 #6049 @markblokpoel
@danielskatz I've added community guidelines based on your suggestions, thank you so much for pointing me in the right direction.

I've addressed the other concerns as best as I could. Thanks again to @stephenfmann for their elaborate feedback and concrete suggestions! Please let me know if everything is in order.

One comment that is open is a discussion of the state of the field. There is to my knowledge no other software infrastructure available that serves this exact purpose.

mathlib is designed to facilitate implementing computational-level simulations for fully specified formal theories (cf. Blokpoel & van Rooij, 2021). While computational modeling is quite widespread in cognitive science, the approach  to computational theory laid out by Blokpoel and van Rooij is, unfortunately, still rare.

 I'm happy to write about this in the paper, but it felt beyond the scope of the paper for JOSS.
