### 2024-05-23 #6049 @danielskatz
> One comment that is open is a discussion of the state of the field. There is to my knowledge no other software infrastructure available that serves this exact purpose.
>
> mathlib is designed to facilitate implementing computational-level simulations for fully specified formal theories (cf. Blokpoel & van Rooij, 2021). While computational modeling is quite widespread in cognitive science, the approach to computational theory laid out by Blokpoel and van Rooij is, unfortunately, still rare.

You might just say this in the paper (if you don't already...)
