### 2024-05-23 #6049 @larkz
Hi gentleman, I had chance to review the updated document and code. Thank you for being patient.

There is a noticeable improvement in the writing quality, since the last manuscript. The clarity of the provided examples facilitates understanding the library's objectives effectively. The presentation on the example Notebooks are quite nice, and easy to follow. Upon executing the code, it seems to be running fine, though some points warrant suggestions and clarifications.

Nevertheless, I would suggest these items to be clarified/resolved before moving forward with the final version for publication.

**Questions/Suggestions:**

- Lines 41-42: claiming to the goal to be writing sustainable code and theoretical contribution appears somewhat self-evident in any software or research. What is new here?

- Lines 43-45: the emphasis on backwards compatibility aligns with common software engineering principles. Perhaps, focusing on `mathlib`'s specific approach to achieving rather than general discussion.

- Lines 51-56: could benefit from a comparative analysis with alternative programming languages. A direct head-to-head comparison, perhaps with Python or Julia, would provide a better understanding of Scala's usefulness in formal theory construction, improving the paper's positioning and persuasiveness.

- Line 76: what are the key differences between "coherence" and "SAT solvers"? This would hwlp readers' understanding - especially for those of us not experts in the field. I'm reading this text and I feel there are some strong similarities.

- Line 86: Are C^+ and C^- disjoint sets? Could they also be not disjoint? Or is this a must-have?

- Line 78: Typo - "to belief ..." vs "to believe"

- Line 89: Are variables A and R are defined, they don't seem to be used in the notebook example.

**General Comments:**

- The paper's primary contribution lies in its innovative verification tools (lines 93-94), which could be more prominently emphasized to highlight the contribution.

- So only the `coherence.ipynb` is available for experimentation. Could we have at least 1-2 more examples fully completed in this same fashion as `coherence.ipynb` before JOSS release?

- Question about the theory. Could there be multiple truth assignments that are somewhat coherent to different degrees? Could there be a preference ranking over them? That is, some are more coherent than others? So it is not just satistfibility, rather also optimization.

- Creating an SBT configuration/entry point, would streamline the process for interested readers who are already familiar with Scala. i.e. `sbt run coherence`

- What is the scalability of `mathlib` to accommodate larger truth tables or complex graphs? This could be useful for readers who are interested in working with this platform further to develop advanced solvers.
