### 2024-05-23 #6049 @danielskatz
@larkz - thanks for your updated comments.  Looking at [your checklist](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-1848383214), I wonder if you've had a chance to look at the software as well?
