### 2024-05-24 #6049 @larkz
> @larkz - thanks for your updated comments. Looking at [your checklist](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-1848383214), I wonder if you've had a chance to look at the software as well?

hi @danielskatz , I've checked off the boxes that I feel have been sufficed by this iteration, there are still some things that I would desire, such as,

- More than just one example.
- Performance guarantees/analysis
- Automated testing, (ie unit tests, SBT entry points etc.)

Which are reflected on the checklist.
