### 2024-05-24 #6049 @danielskatz
HI @larkz - thanks for this update.

From the point-of-view of JOSS and the [review criteria](https://joss.readthedocs.io/en/latest/review_checklist.html), I think it's reasonable to ask for a second example. However, the other two items you ask for might be beyond what JOSS strictly requires.

For performance, we say `Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)` so unless you strongly feel that performance is a requirement for this specific software while it isn't for JOSS-published software in general, I think I would treat your comment as a future feature request that than something that would stop the JOSS review.

And similarly, for automated tests, we say `Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?` so while we would prefer automated tests, they are not required if there is at least a manual testing process.  Also, see https://joss.readthedocs.io/en/latest/review_criteria.html#tests

Given this, I would expect that @markblokpoel should be able to add another example.  And I would ask you to reconsider if the software satisfies the JOSS criteria as written, even though it may not be as advanced as it could be.

thanks!
