### 2024-05-24 #6049 @drussellmrichie
As mentioned, I am not a Scala user, so I am not qualified to comment on the functionality or tests of this package. But I did look through the repo and the paper to evaluate other components as best I could. Some edits to the paper:

Line 20: "to facilitate users to implement". I would rewrite as either: "to help users implement" or "to facilitate users' implementations of"

Line 27: "supports scholars to write". I would rewrite as: "helps scholars write" or "supports scholars in writing".

Line 40: same as line 27

Lines 78 and 79: "to belief" -> "to believe"

Line 89: What are A and R? Suspect input to Coh+ should just be T, especially given table 2.

I will check off "Quality of writing" once these are accepted or rejected. I will also check off "State of the field" once that is completed. I second @danielskatz 's suggestion to say that there are no comparable software packages (if that is indeed the case).

Last, for what it's worth, I'm not sure I'm sold on the readability argument. I look at this Scala code, and it doesn't look much more readable to me than Python! Am I missing something? Can an argument be made in the article about why this is more readable and verifiable than Python?

As a reader I might also like to see actually executed/simulated examples, but maybe that is beyond the scope of the paper.
