### 2024-05-31 edae705
State of field statement added
