### 2024-05-31 5d8a773
Illustration non-functional vs functional code
