+++ paper.md
@@ -25,7 +25,7 @@ set theory, functions, first-order logic, graph theory, etc.) allow
 scientists to 'conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined'
 [@guest_martin:2021]. They make otherwise underspecified theories precise and open for critical reflection
 [@vanrooij_baggio:2021]. A theory can be formally specified in a computational model using mathematical concepts
-such as set theory, graph theory and probability theory.  The specification is often followed by analysis to understand
+such as set theory, graph theory and probability theory. The specification is often followed by analysis to understand
 precisely what assumptions and consequences the formal theory entails. An important method of analysis is computer
 simulation, which allows scientists to explore complex model behaviours and derive predictions that otherwise cannot
 be analytically derived[^1].
@@ -33,13 +33,13 @@ be analytically derived[^1].
 [^1]: Computer simulations can also aid scientists to discover properties of the model that they would not have
 thought to analytically derive, even when in principle the property can be analytically derived.
 
-```mathlib``` is a library for Scala [@odersky:2008] supporting functional programming that resembles mathematical 
+```mathlib``` is a library for Scala [@odersky:2008] supporting functional programming that resembles mathematical
 expressions such as set theory and graph theory. This library was developed to complement the
 theory development method outlined in the open education book Theoretical modeling for cognitive science and
 psychology by @blokpoel_vanrooij:2021. To date ```mathlib``` is the only library that facilitates implementing
 computational-level simulations for fully specified formal theories.
 
-The goal of this library is to facilitate users to implement simulations of their formal theories. Code written in
+The goal of this library is to help users to implement simulations of their formal theories. Code written in
 Scala using ```mathlib``` is:
 
 * easy to **read**, because ```mathlib``` syntax closely resembles mathematical notation
@@ -48,20 +48,53 @@ Scala using ```mathlib``` is:
 
 # Statement of need
 
-```mathlib``` supports scholars to write **readable** and **verifiable** code. Writing code is not easy, writing
+```mathlib``` supports scholars in writing **readable** and **verifiable** code. Writing code is not easy, writing
 code for which we can know that it computes what the specification (i.e., the formal theory) states is even harder.
 Given the critical role of theory and computer simulations in cognitive science, it is important that scholars
 can verify that the code does what the authors intend it to do. This can be facilitated by having a programming language
-where the
-syntax and semantics closely matches that of the specification. Since formal theories are specified using mathematical
+where the syntax and semantics closely matches that of the specification. Since formal theories are specified using
+mathematical
 notation [@marr:1982; @blokpoel_vanrooij:2021; @guest_martin:2021], functional programming languages bring a lot to the
 table in terms of syntactic and semantic resemblance to mathematical concepts and notation. ```mathlib``` adds
 mathematical concepts and notation to the functional programming language Scala [@odersky:2008]. At the time of writing,
-the current version (0.9.1) supports set theory and graph theory. In the next section, we provide two examples to
-illustrate how Scala and ```mathlib``` make code more accessible to verify the relationship between simulation and
-theory.
[…]
