### 2024-05-31 7960432
Illustration non-functional vs functional code
