### 2024-05-31 237696c
Updated demo instructions in README.md
