### 2024-05-31 #6049 @markblokpoel
:pray::pray: A huge thanks to all the reviewers @stephenfmann, @larkz and @drussellmrichie and the editor @danielskatz for your constructive feedback. It has grealy improved the paper and software. :pray::pray:

## Below are my responsed to @larkz's [review](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2127103359).

*Lines 41-42:* claiming to the goal to be writing sustainable code and theoretical contribution appears somewhat self-evident in any software or research. What is new here?"

*Lines 43-45:* the emphasis on backwards compatibility aligns with common software engineering principles. Perhaps, focusing on mathlib's specific approach to achieving rather than general discussion.

*Reply:* These this notion and goals might be self-evident in computer science, but is not in cognitive science (the target audience of this library). My aim with mathlib is to make these goals more accessible while interfacing with a specific notion of formal theory, as explained in the paper lines 31-36 and illustrated in the two examples.

*Lines 51-56:* could benefit from a comparative analysis with alternative programming languages. A direct head-to-head comparison, perhaps with Python or Julia, would provide a better understanding of Scala's usefulness in formal theory construction, improving the paper's positioning and persuasiveness.

*Reply:* I do not wish to take a normative stance in my paper. I present mathlib relative to its own goals and wish to acknowledge that other routes to the same goals are possible. A comparison may only give rise to a 'this is better than that' debate which I do not wish to encourage.

*Line 76:* what are the key differences between "coherence" and "SAT solvers"? This would hwlp readers' understanding - especially for those of us not experts in the field. I'm reading this text and I feel there are some strong similarities.

*Reply:* While there are certainly commonalities between coherence theory and SAT solvers, there are important theoretical differences. I do not think that this paper is the place to discuss these non-code related topics and hence I refer the reader to the literature on this debate. Thagard & Verbeurgt (1998) explain in their paper the theoretical aims for Coherence theory, which is to explain, not to solve (cf. Cummins, 2000; Blokpoel and van Rooij 2021).

*Line 86:* Are C^+ and C^- disjoint sets? Could they also be not disjoint? Or is this a must-have?

*Reply:* Constraints between beliefs are either positive or negative and cannot be both at the same time. This is a theoretical commitment by the original authors and I faithfully represent their theory in this paper.

*Line 78:* Typo - "to belief ..." vs "to believe"

*Reply:* Fixed, thank you!

*Line 89:* Are variables A and R are defined, they don't seem to be used in the notebook example.

*Reply:* I believe this was fixed in an earlier revision based on @stephenfmann's review.

*General comments:*

The paper's primary contribution lies in its innovative verification tools (lines 93-94), which could be more prominently emphasized to highlight the contribution.

*Reply:* Thank you for acknowledging the primary contribution. I believe that the line references might be of an earlier version of the paper. In the current version, the contribution is explained in lines 29-35 and with the two examples.

So only the coherence.ipynb is available for experimentation. Could we have at least 1-2 more examples fully completed in this same fashion as coherence.ipynb before JOSS release?

*Reply:* There are two demos available in the mathlib repository: `SubsetChoice.scala` and `Coherence.scala`. While I'm happy to hear about all reviewer's appreciation for the notebook tutorials, these are a separate academic contribution (as discussed [here](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2004659999 and supported by @danielkatz [here](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-1881429073)). I believe the two demos satisfy the conditions for publication in JOSS.

Question about the theory. Could there be multiple truth assignments that are somewhat coherent to different degrees? Could there be a preference ranking over them? That is, some are more coherent than others? So it is not just satistfibility, rather also optimization.

*Reply:* These are a very good observations and ones that we often discuss in our class where we teach formal modelling (Blokpoel and van Rooij, 2021). The theory postulates that humans optimize Coherence scores (though in TV98 the authors also propose an approximation strategy at Marr's (1998) algorithmic-level of analysis). The 'ranking' is in a sense defined by the coherence score. I'm happy to clarify this, but these questions fall outside the scope of the current paper and have been discussed in other publications (e.g. Blokpoel & van Rooij, 2021 Chapter 5; Thagard and Verbeurgt, 1998; van Rooij & Wareham, 2012).

van Rooij, I., & Wareham, T. (2012). Intractability and approximation of optimization theories of cognition. Journal of Mathematical Psychology, 56(4), 232-247.

Creating an SBT configuration/entry point, would streamline the process for interested readers who are already familiar with Scala. i.e. sbt run coherence

*Reply:* A great suggestion, I've included this in the `build.sbt` file and installation instructions in the readme and on the website.

What is the scalability of mathlib to accommodate larger truth tables or complex graphs? This could be useful for readers who are interested in working with this platform further to develop advanced solvers.

*Reply:* As mentioned by @danielkatz [here](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2129393700) and since I make clear here that the focus of mathlib is not about performance (lines 54-57), this analysis will not be included in the paper.

## Below is my reply to the [final open point](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2134664061) by @stephenfmann:

One of the checklist comments for the paper is "State of the field: Do the authors describe how this software compares to other commonly-used packages?" There's a discussion of Scala more generally but there don't appear to be other commonly-used packages for this specific use case. I'll leave it unchecked until the author confirms this is true.

*Reply:* I've added a clarification to the manuscript (lines 19-20) to clarify that there is to my knowledge no other software infrastructure available that serves the exact purpose of implementing computational-level simulations for fully specified formal theories (cf. Blokpoel & van Rooij, 2021).

## Below are my replies @drussellmrichie's [review](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2130005541):

Thank you for the feedback on the writing. I've updated the manuscript accordingly. However, the final comment on A and R seems to be from an older version of the manuscript. I believe that was a copy-paste mistake on my part. The current version does not mention sets A and R.

Last, for what it's worth, I'm not sure I'm sold on the readability argument. I look at this Scala code, and it doesn't look much more readable to me than Python! Am I missing something? Can an argument be made in the article about why this is more readable and verifiable than Python?

*Reply:* As explained in my reply to @larkz's review: I do not wish to take a normative stance in my paper. I present mathlib relative to its own goals and wish to acknowledge that other routes to the same goals are possible. A comparison may only give rise to a 'this is better than that' debate which I do not wish to encourage.

I do acknowledge that, especially to novel Scala users, the code might not be easily readable. This is also not realistic to expect nor do I make this claim in my paper. The point is that the functional programming paradigm
allows for easier to prove of code semantics (i.e., that the code implements the formal specification) compared to non-functional programming. This full discussion seems to be out of scope for the current submission,
but I've added a clarifying illustration (lines 39-47).

As a reader I might also like to see actually executed/simulated examples, but maybe that is beyond the scope of the paper.

*Reply:* There are two example codes accompanying the illustrations in the paper. These are, intentionally, not elaborate simulations as I agree that this is beyond the scope of the paper.
