### 2024-06-03 #6049 @danielskatz
@larkz & @drussellmrichie - can you take a look at @markblokpoel's [responses](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2141580877) and see if you can check off more items on your list? Or respond with what else you think is needed. (Sorry if I've lost track of something else, here, and if so, please remind me.)
