### 2024-06-03 #6049 @drussellmrichie
I've checked off all the boxes that I am qualified to check off (again, I don't know Scala and so am unable to check eg the installation and execution of the package).

I did find a new typo:

Line 44: "mode" should be "more" I believe.
