### 2024-06-15 #6049 @danielskatz
👋 @larkz - can you take a look at @markblokpoel's https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2141580877 and other comments above, and see if you can check off more items on your list? Or respond with what else you think is needed. (Sorry if I've lost track of something else, here, and if so, please remind me.)
