### 2024-06-24 #6049 @markblokpoel
Regarding the additional notebook example, please see my reply [here](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2141580877). (Copy pasted below for convenience.) @danielskatz given your previous decision on this point, I will leave this up to you. However, if the notebooks will be considered part of this submission, it will be a considerable effort and new intellectual contributions to make those ready for publication.

> There are two demos available in the mathlib repository: SubsetChoice.scala and Coherence.scala. While I'm happy to hear about all reviewer's appreciation for the notebook tutorials, these are a separate academic contribution (as discussed https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2004659999 and supported by @danielkatz https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-1881429073). I believe the two demos satisfy the conditions for publication in JOSS.

Regarding state of the field, I have addressed this in response to @stephenfmann's review, see my reply [here](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2141580877). (Copy pasted below for convenience.) @danielskatz please see if the additions to the manuscript on lines 19-20 suffice.

>  I've added a clarification to the manuscript (lines 19-20) to clarify that there is to my knowledge no other software infrastructure available that serves the exact purpose of implementing computational-level simulations for fully specified formal theories (cf. Blokpoel & van Rooij, 2021).
