### 2024-07-04 #6049 @larkz
Thank you all for your patience, in waiting for my feedback.

I am here to support @markbloepel in enhancing the paper and code's credibility and quality.

Currently, I believe the previous response could be more thorough. For instance, `sympy` in Python, `algebird` in Scala, and `containers` in Haskell offer some functionalities similar to this current software package. Of course, the author knows exactly how their software package differs from these, but it would be even better if this was clearly communicated.

I also think that providing only one example is not the most ideal. The requirements clearly ask for plurality of examples, and here there is just one. So if time permits, I would really appreciate a second detailed example to strengthen the presentation of the code.

However, if there is any urgency to proceed, please do not let my suggestions hold anyone back. My role is simply to provide an honest and independent assessment, and I am not rigidly against relaxing some of the requirements.

@danielskatz please let me know what your expectations are. I feel some minor revisions/updates should be made, but if you feel this codebase and presentation its sufficient as is, as mentioned earlier, I won’t hold anyone back here. 🙂
