### 2024-07-12 035a50f
Update paper.md
Acknowledgements update.
