+++ paper.md
@@ -284,8 +284,8 @@ def coherence(
 # Acknowledgements
 
 We thank the Computational Cognitive Science group at the Donders Center for Cognition (Nijmegen, The Netherlands)
-for useful discussions and feedback, in particular, Laura van de Braak, Olivia Guest and Iris van Rooij. We also thank
-Larkin Liu for their useful feedback and review.
+for useful discussions and feedback, in particular, Laura van de Braak, Olivia Guest and Iris van Rooij. We also thank the reviewers
+Larkin Liu (@larkz), Russel Richie (@drussellmrichie) and Stephen Mann (@stephenfmann) for their useful feedback which has greatly improved this paper.
 
 This project was supported by Netherlands Organization for Scientific Research Gravitation Grant of the
 Language in Interaction consortium 024.001.006, the Radboud School for Artificial Intelligence and the
