### 2024-07-12 ed3d4f7
Update paper.md
