### 2024-07-12 17861cf
Update paper.md
Typo
