### 2024-07-12 #6049 @markblokpoel
Hi @larkz,

Thanks again for the feedback, I appreciate it greatly!

> Currently, I believe the previous response could be more thorough. For instance, sympy in Python, algebird in Scala, and containers in Haskell offer some functionalities similar to this current software package. Of course, the author knows exactly how their software package differs from these, but it would be even better if this was clearly communicated.

In the manuscript (lines 19-21) I clarify that there is to my knowledge no other software infrastructure available that serves the exact purpose of implementing computational-level simulations for fully specified formal theories (cf. Blokpoel & van Rooij, 2021). Other approaches to set theory and graph theory exist, and while these are functionally far more advanced than MathLib, they (a) do not match the method from the book and (b) due to them being advanced are not accessible to a wider cognitive science field.

Perhaps you missed these lines in the many updates there have been, but @larkz I think you can check off the text box for this.

> I also think that providing only one example is not the most ideal. The requirements clearly ask for plurality of examples, and here there is just one. So if time permits, I would really appreciate a second detailed example to strengthen the presentation of the code.

Like I mentioned [before](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2141580877), there are two demos available in the mathlib repository: SubsetChoice.scala and Coherence.scala. While I'm happy to hear about all reviewer's appreciation for the notebook tutorials, these are a separate academic contribution (as discussed https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2004659999 and supported by @danielkatz https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-1881429073). I believe the two demos satisfy the conditions for publication in JOSS.

Hi @danielskatz, I think given the above, the manuscript is ready to be published. Thank you all so much for the help.
