### 2024-07-12 #6049 @danielskatz
@larkz - can you now check off the final three items on [your checklist](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-1848383214)?
