### 2024-07-13 #6049 @larkz
Just some technical notes for @markblokpoel.

In the paper it says, `Other approaches to set theory and graph theory exist,`, what exactly are these packages? I think other people who worked on similar software should be at least be acknowledged. Could you cite a few similar works perhaps?

I also suggest to @markblokpoel to add more examples to this repo in the future. Because as it stands, the repo itself is kind of weak in terms of descriptive and runnable examples.

So I will check off these boxes as @danielskatz requested. Please acknowledge this is under a relaxation of what is strictly required from the checklist - at least from my point of view. Thank you for all your hard work, and best of luck 🙁 .
