### 2024-07-23 #6049 @markblokpoel
:wave: @lark - thank you for your work in reviewing the submission and your valuable feedback. Based on my [previous comments](https://github.com/openjournals/joss-reviews/issues/6049#issuecomment-2225964036), I think this is now complete.
