### 2024-07-23 a2dadd2
minor changes in paper.bib
