+++ paper.bib
@@ -17,7 +17,7 @@
 }
 
 @article{thagard:1998,
-    title = {Coherence as {{Constraint Satisfaction}}},
+    title = {Coherence as Constraint Satisfaction},
     author = {Thagard, Paul and Verbeurgt, Karsten},
     year = {1998},
     journal = {Cognitive Science},
@@ -48,7 +48,7 @@
 
 @book{odersky:2008,
     author = {{Odersky}, M.},
-    title = {Programming in Scala.},
+    title = {Programming in {S}cala},
     year = 2008,
     publisher = {Mountain View, California: Artima}
 }
