### 2024-07-25 #6049 @markblokpoel
@danielskatz Here are the items:

- The tagged release: [https://github.com/markblokpoel/mathlib/releases/tag/v0.9.1](https://github.com/markblokpoel/mathlib/releases/tag/v0.9.1)
- The Zenodo deposit: [https://zenodo.org/records/12819230](https://zenodo.org/records/12819230)
- The DOI: [10.5281/zenodo.12819230](https://doi.org/10.5281/zenodo.12819230)

Hope I did this correctly, please let me know if the release and deposit are in order.
