### 2024-07-25 6fda9f0
Update paper.bib with DOIs
