+++ paper.bib
@@ -25,7 +25,8 @@
     number = {1},
     pages = {24},
     langid = {english},
-    keywords = {Coherence}
+    keywords = {Coherence},
+    doi = {10.1207/s15516709cog2201_1}
 }
 
 @article{vanrooij_baggio:2021,
@@ -43,7 +44,8 @@
     author = {{Marr}, D.},
     title = {Vision: A computational investigation into the human representation and processing of visual information},
     year = 1982,
-    publisher = {W.H. Freeman, San Francisco, CA}
+    publisher = {W.H. Freeman, San Francisco, CA},
+    doi = {10.7551/mitpress/9780262514620.001.0001}
 }
 
 @book{odersky:2008,
