### 2024-07-25 #6049 @danielskatz
@markblokpoel - the DOI seems to be having some problems, so I'm going to leave this open until it's working.  In the meantime, however:

Congratulations on your publication!!

And thanks to @larkz, @stephenfmann, and @drussellmrichie for reviewing!
JOSS depends on volunteers and couldn't succeed without you
