### 2024-07-25 #6049 @danielskatz
Let's see if there was just a transient error, since Crossref (where the DOIs are registered) has [all green status](https://status.crossref.org) currently.
