### 2024-07-25 #6049 @danielskatz
Ok, this DOI is now registered and working, so I'm going to close this issue and end this process - thanks all again!!
